requirements.txt
asgiref==3.8.1
decouple==0.0.7
Django==5.1.3
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
PyJWT==2.10.0
python-decouple==3.8
sqlparse==0.5.2
tzdata==2024.2
